# docs/swagger.yaml
definitions:
...
  controller.deleteUserInput:
    properties:
      password:
        type: string
    type: object
...
  controller.specificationInput:
    properties:
      additional_feature:
        type: string
      battery:
        type: string
      camera:
        type: integer
      memory:
        type: integer
      network:
        type: string
      operating_system:
        type: string
      storage:
        type: integer
    required:
    - battery
    - camera
    - memory
    - network
    - operating_system
    - storage
    type: object
  controller.userUpdate:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
...
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      review_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Phone:
    properties:
      brand_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      model:
        type: string
      price:
        type: integer
      release_date:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      specification:
        items:
          $ref: '#/definitions/models.Specification'
        type: array
      updated_at:
        type: string
    type: object
  models.Profile:
    properties:
      biodata:
        type: string
      birthday:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: integer
      image_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Review:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      phone_id:
        type: integer
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Specification:
    properties:
      additional_feature:
        type: string
      battery:
        type: string
      camera:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      memory:
        type: integer
      network:
        type: string
      operating_system:
        type: string
      phone_id:
        type: integer
      storage:
        type: integer
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      profiles:
        items:
          $ref: '#/definitions/models.Profile'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /phones/{id}/specification:
    get:
      description: Get all Phones data by phone id.
      parameters:
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Phone'
            type: array
      summary: Get specification data by Phone id. (PUBLIC)
      tags:
      - Phones
    post:
      description: Creating a specification data for phone, only admin can access
        this route
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      - description: example JSON body to create a specification for phone
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.specificationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Specification'
      security:
      - BearerToken: []
      summary: Create Specification for phone (ADMIN ONLY)
      tags:
      - Phones
    put:
      description: Creating a specification data for phone, only admin can access
        this route
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: phone id
        in: path
        name: id
        required: true
        type: string
      - description: example JSON body to update a specification for phone
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.specificationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Specification'
      security:
      - BearerToken: []
      summary: Update Specification for phone (ADMIN ONLY)
      tags:
      - Phones
...
  /users:
    delete:
      description: Will delete the user account itself, user ID is taken from JWT
        Token so only acount's owner can delete its own accout
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to delete user's own account
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.deleteUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - BearerToken: []
      summary: Delete user's own account
      tags:
      - Users
    get:
      description: Get a list of account with 'user' role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: Get all account with user role. (PUBLIC)
      tags:
      - Users
  /users/:
    put:
      description: update its own user data, user ID is taken from JWT Token so only
        acount's owner can update the user information
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: tExample JSON body to update User data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.userUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerToken: []
      summary: Update User data.
      tags:
      - Users
  /users/{id}:
    delete:
      description: Delete a User by id, only admin can access this route
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - BearerToken: []
      summary: Delete User by id (ADMIN ONLY)
      tags:
      - Users
    get:
      description: Get user data by ID.
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: Get single user by ID. (PUBLIC)
      tags:
      - Users
  /users/{id}/profile:
    get:
      description: Get all Users profile data by user id.this will only display account
        with role "user", if admin not create profile yet the profile will not appear
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Get profiles data by User id. (PUBLIC)
      tags:
      - Users
swagger: "2.0"